+++ web/lang/en.php
@@ -167,6 +167,7 @@ $translations = array(
     'labels.weekstart' => 'Week starts on',
     'labels.show_week_nb' => 'Show week numbers',
     'labels.show_now_indicator' => 'Show a marker indicating the current time',
+    'labels.list_days' => 'List view days',
 
 
 // Messages
@@ -211,6 +212,7 @@ $translations = array(
 
     'messages.help_defaultcalendar' => 'New events will be placed in this calendar by default. Set here your most used calendar',
     'messages.help_timezone' => 'Choose your current timezone. This will affect how you see existing events, and new events will be created using this timezone',
+    'messages.help_list_days' => 'Choose how many consecutive days do you want to be shown on the "list" view',
 
     'messages.info_prefssaved' => 'Preferences saved',